alpine => debian-slim
It seems that on amd64 with qemu you might be able to build for alpine arm32, but you can't run it.
But I have no way to test if the build even works then. So, no alpine for now.

arm32v6 emulation runtime error:
arm-binfmt-P: Could not open '/lib/ld-linux-armhf.so.3': No such file or directory